django_api/app/requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
flake8==4.0.1
django-split-settings==1.1.0
django-debug-toolbar==3.2.4
python-dotenv==0.19.2
gunicorn==20.1.0
